+++ lib/src/models/binding.dart
@@ -7,8 +7,8 @@ typedef CactusModel = Pointer<CactusModelOpaque>;
 typedef CactusTokenCallbackNative = Void Function(Pointer<Utf8> token, Uint32 tokenId, Pointer<Void> userData);
 typedef CactusTokenCallbackDart = void Function(Pointer<Utf8> token, int tokenId, Pointer<Void> userData);
 
-typedef CactusInitNative = CactusModel Function(Pointer<Utf8> modelPath, Size contextSize);
-typedef CactusInitDart = CactusModel Function(Pointer<Utf8> modelPath, int contextSize);
+typedef CactusInitNative = CactusModel Function(Pointer<Utf8> modelPath, Size contextSize, Pointer<Utf8> corpusDir);
+typedef CactusInitDart = CactusModel Function(Pointer<Utf8> modelPath, int contextSize, Pointer<Utf8> corpusDir);
 
 typedef CactusCompleteNative = Int32 Function(
     CactusModel model,
@@ -32,6 +32,9 @@ typedef CactusCompleteDart = int Function(
 typedef CactusDestroyNative = Void Function(CactusModel model);
 typedef CactusDestroyDart = void Function(CactusModel model);
 
+typedef CactusResetNative = Void Function(CactusModel model);
+typedef CactusResetDart = void Function(CactusModel model);
+
 typedef CactusEmbedNative = Int32 Function(
     CactusModel model,
     Pointer<Utf8> text,
@@ -45,6 +48,29 @@ typedef CactusEmbedDart = int Function(
     int bufferSize,
     Pointer<Size> embeddingDim);
 
+typedef CactusTranscribeNative = Int32 Function(
+    CactusModel model,
+    Pointer<Utf8> audioFilePath,
+    Pointer<Utf8> prompt,
+    Pointer<Utf8> responseBuffer,
+    Size bufferSize,
+    Pointer<Utf8> optionsJson,
+    Pointer<NativeFunction<CactusTokenCallbackNative>> callback,
+    Pointer<Void> userData,
+    Pointer<Uint8> pcmBuffer,
+    Size pcmBufferSize);
+typedef CactusTranscribeDart = int Function(
+    CactusModel model,
+    Pointer<Utf8> audioFilePath,
+    Pointer<Utf8> prompt,
+    Pointer<Utf8> responseBuffer,
+    int bufferSize,
+    Pointer<Utf8> optionsJson,
+    Pointer<NativeFunction<CactusTokenCallbackNative>> callback,
+    Pointer<Void> userData,
+    Pointer<Uint8> pcmBuffer,
+    int pcmBufferSize);
+
 typedef RegisterAppNative = Pointer<Utf8> Function(
     Pointer<Utf8> encData);
 typedef RegisterAppDart = Pointer<Utf8> Function(
@@ -53,57 +79,5 @@ typedef RegisterAppDart = Pointer<Utf8> Function(
 typedef GetAllEntriesNative = Pointer<Utf8> Function();
 typedef GetAllEntriesDart = Pointer<Utf8> Function();
 
-typedef GetDeviceIdNative = Pointer<Utf8> Function();
-typedef GetDeviceIdDart = Pointer<Utf8> Function();
-
-// Whisper model types
-final class WhisperContextOpaque extends Opaque {}
-typedef WhisperContext = Pointer<WhisperContextOpaque>;
-
-final class WhisperStateOpaque extends Opaque {}
-typedef WhisperState = Pointer<WhisperStateOpaque>;
-
-final class WhisperFullParamsOpaque extends Opaque {}
-typedef WhisperFullParams = Pointer<WhisperFullParamsOpaque>;
-
-// Whisper enums
-abstract class WhisperSamplingStrategy {
-  static const int whisperSamplingGreedy = 0;
-  static const int whisperSamplingBeamSearch = 1;
-}
-
-// Whisper function bindings
-typedef WhisperInitFromFileNative = WhisperContext Function(Pointer<Utf8> pathModel);
-typedef WhisperInitFromFileDart = WhisperContext Function(Pointer<Utf8> pathModel);
-
-typedef WhisperFreeNative = Void Function(WhisperContext ctx);
-typedef WhisperFreeDart = void Function(WhisperContext ctx);
-
-typedef WhisperFreeParamsNative = Void Function(WhisperFullParams params);
-typedef WhisperFreeParamsDart = void Function(WhisperFullParams params);
-
-typedef WhisperFullDefaultParamsByRefNative = WhisperFullParams Function(Int32 strategy);
-typedef WhisperFullDefaultParamsByRefDart = WhisperFullParams Function(int strategy);
-
-typedef WhisperFullNative = Int32 Function(
-    WhisperContext ctx,
-    WhisperFullParams params,
-    Pointer<Float> samples,
-    Int32 nSamples);
-typedef WhisperFullDart = int Function(
-    WhisperContext ctx,
-    WhisperFullParams params,
-    Pointer<Float> samples,
-    int nSamples);
-
-typedef WhisperFullNSegmentsNative = Int32 Function(WhisperContext ctx);
-typedef WhisperFullNSegmentsDart = int Function(WhisperContext ctx);
-
-typedef WhisperFullGetSegmentTextNative = Pointer<Utf8> Function(WhisperContext ctx, Int32 iSegment);
-typedef WhisperFullGetSegmentTextDart = Pointer<Utf8> Function(WhisperContext ctx, int iSegment);
-
-typedef WhisperFullGetSegmentT0Native = Int64 Function(WhisperContext ctx, Int32 iSegment);
-typedef WhisperFullGetSegmentT0Dart = int Function(WhisperContext ctx, int iSegment);
-
-typedef WhisperFullGetSegmentT1Native = Int64 Function(WhisperContext ctx, Int32 iSegment);
-typedef WhisperFullGetSegmentT1Dart = int Function(WhisperContext ctx, int iSegment);
+typedef GetDeviceIdNative = Pointer<Utf8> Function(Pointer<Utf8> token);
+typedef GetDeviceIdDart = Pointer<Utf8> Function(Pointer<Utf8> token);